==> code/0.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

no_ = sorted(
    [na for na in os.listdir() if re.match(r"^[^._].*\.ipynb$", na)],
    key=lambda na: int(na.split(".")[0]),
)

no_.remove("0.ipynb")

for no in no_:

    kwat.shell.run(
        "jupyter nbconvert --execute --ExecutePreprocessor.timeout=-1 --inplace {}".format(
            no
        )
    )
