==> code/1_process_feature_by_sample.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

nu_fe_sa = pd.read_csv("{}number_feature_sample.tsv".format(PAI), sep="\t", index_col=0)

nu_fe_sa.columns.name = "Sample"

nu_fe_sa

if 0 < len(SETTING["pr"]):

    nu_fe_sa = kwat.feature_by_sample.process(nu_fe_sa, **SETTING["pr"])

nu_fe_sa.to_csv(path_or_buf="{}number_feature_sample.tsv".format(PAO), sep="\t")

kwat.plot.plot_histogram(
    [se for co, se in nu_fe_sa.iteritems()],
)

nu_fe_sa.loc[SETTING["fe_"], :]
