==> code/5_gsea.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

bi_co_sa = pd.read_csv("../output/01_comparison_sample.tsv", sep="\t", index_col=0)

se_fe_ = kwat.gmt.read(kwat.path.list("../input/set/"))

st_se_in = pd.read_csv("../output/statistic_set_information.tsv", sep="\t", index_col=0)

for co in bi_co_sa.index:

    pa = "{}/{}".format(kwat.path.clean(co), SETTING["sc"])

    sc_ = pd.read_csv(
        "../output/compare_feature/{}/statistic.tsv".format(pa),
        sep="\t",
        index_col=0,
    ).loc[:, "Score"]

    pa = "../output/score_set/{}/".format(pa)

    kwat.path.make(pa)

    kwat.score_set.score_samples_and_sets(
        sc_.to_frame(),
        {se: se_fe_[se] for se in st_se_in.index[st_se_in.loc[:, "Selected"]]},
        pa="{}statistic.tsv".format(pa),
    )

    for se in SETTING["se_"]:

        kwat.score_set.score_sample_and_set(
            sc_,
            se_fe_[se],
            title="{}<br>{}".format(co, se),
            pa="{}{}.html".format(pa, kwat.path.clean(se)),
        )
