==> code/6_make_set_by_information.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

bi_co_sa = pd.read_csv("{}01_comparison_sample.tsv".format(PAO), sep="\t", index_col=0)

nu_fe_sa = pd.read_csv("{}number_feature_sample.tsv".format(PAO), sep="\t", index_col=0)

se_fe_ = read_set()

for se in se_fe_:

    if re.search(
        r"^{0}$|.*{1}{0}{1}.*|^{0}{1}.*|.*{1}{0}$".format("FATIGUE[0-9]*", "[\s_]"),
        se,
    ):

        print(se)

fe__ = se_fe_.values()

fe_no = {fe: None for fe in nu_fe_sa.index.values}

st_se_in = pd.DataFrame(
    data={
        "Selected": False,
        "N": [len(fe_) for fe_ in fe__],
        "n": [sum(fe in fe_no for fe in fe_) for fe_ in fe__],
    },
    index=se_fe_,
)

st_se_in.loc[:, "."] = st_se_in.loc[:, "n"].values / st_se_in.loc[:, "N"].values

st_se_in

st_se_in.loc[
    kwat.array.check_is_extreme(
        st_se_in.loc[:, "n"].values,
        ">",
        th_=[None, SETTING["n_ge"]],
    )
    & kwat.array.check_is_extreme(
        st_se_in.loc[:, "."].values,
        ">",
        th_=[None, SETTING["._ge"]],
    ),
    "Selected",
] = True

assert (1 / 2 < st_se_in.loc[SETTING["se_"], "."]).all(), st_se_in.loc[
    SETTING["se_"], "."
]

st_se_in.loc[SETTING["se_"], "Selected"] = True

st_se_in.to_csv(path_or_buf="{}statistic_set_information.tsv".format(PAO), sep="\t")

print(st_se_in.loc[:, "Selected"].sum())

st_se_in

st_se_in.loc[SETTING["se_"], :].sort_values(".")
