==> code/7.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

bi_co_sa = pd.read_csv("../output/01_comparison_sample.tsv", sep="\t", index_col=0)

sc_se_sa = pd.read_csv("../output/score_set_sample.tsv", sep="\t", index_col=0)

sc_se_sa.isna().sum(1).sort_values()

na_se_fe_ = {}

pa = "../input/set/"

for na in os.listdir(pa):

    na_se_fe_[na] = kwat.gmt.read(["{}{}".format(pa, na)])

fu = kwat.information.get_ic

funa = fu.__name__

ke = {
    "tyta": "binary",
    "st": SETTING["st"],
}

for co, ta in bi_co_sa.iterrows():

    ta.dropna(inplace=True)

    pa = "../output/compare_set/{}/{}/".format(kwat.path.clean(co), funa)

    kwat.path.make(pa)

    st = kwat.function_heat_map.make(
        ta,
        sc_se_sa,
        fu,
        n_jo=SETTING["n_jo"],
        n_sa=SETTING["n_sa"],
        n_sh=SETTING["n_sh"],
        n_pl=SETTING["n_ex"],
        title="All ({})".format(funa),
        pa=pa,
        **ke,
    )

    if 0 < len(SETTING["se_"]):

        kwat.function_heat_map.make(
            ta,
            sc_se_sa.reindex(SETTING["se_"]),
            st,
            n_pl=None,
            title="Peek ({})".format(funa),
            **ke,
        )

        plot_peek(st["Score"], SETTING["se_"], pa)

    for na, se_fe_ in na_se_fe_.items():

        pa2 = "{}{}/".format(pa, na)

        kwat.path.make(pa2)

        se_ = sc_se_sa.index.intersection(se_fe_)

        if 0 < len(se_):

            kwat.function_heat_map.make(
                ta,
                sc_se_sa.loc[se_, :],
                st,
                n_pl=SETTING["n_ex"],
                title="{} ({})".format(na, funa),
                pa=pa2,
                **ke,
            )
