==> code/7_run_prerank_gsea.ipynb <==
%load_ext autoreload
%autoreload 2

import gsea
from __init__ import *

bi_co_sa = pd.read_csv("{}01_comparison_sample.tsv".format(PAO), sep="\t", index_col=0)

se_fe_ = read_set()

st_se_in = pd.read_csv(
    "{}statistic_set_information.tsv".format(PAO), sep="\t", index_col=0
)

for co in bi_co_sa.index:

    pa = "{}/{}".format(kwat.path.clean(co), SETTING["sc"])

    sc_ = pd.read_csv(
        "{}compare_feature/{}/statistic.tsv".format(PAO, pa),
        sep="\t",
        index_col=0,
    ).loc[:, "Score"]

    pa = "{}gsea/{}/".format(PAO, pa)

    kwat.path.make(pa)

    gsea.run_prerank_gsea(
        sc_,
        {se: se_fe_[se] for se in st_se_in.index[st_se_in.loc[:, "Selected"]]},
        mi=None,
        ma=None,
        n_jo=SETTING["n_jo"],
        n_pe=SETTING["n_sh"],
        pa=pa,
    )

    for se in SETTING["se_"]:

        break

        gsea._score_1_1(
            sc_,
            se_fe_[se],
            title="{}<br>{}".format(co, se),
            pa=pa,
        )
