==> code/9_compare_set.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

bi_co_sa = pd.read_csv("{}01_comparison_sample.tsv".format(PAO), sep="\t", index_col=0)

en_se_sa = pd.read_csv("{}enrichment_set_sample.tsv".format(PAO), sep="\t", index_col=0)

bu_se_fe_ = {}

pa = "../input/set/"

for pas in kwat.path.list_directory(pa, ke_=(r".gmt$", r".tsv$")):

    bu = os.path.splitext(os.path.split(pas)[1])[0]

    print(bu)

    bu_se_fe_[bu] = kwat.gmt.read([pas])

fu = kwat.information.get_ic

fun = fu.__name__

ke = {
    "tyt": "binary",
    "st": SETTING["st"],
}

for co, ta in bi_co_sa.iterrows():

    ta.dropna(inplace=True)

    paf = "{}compare_set/{}/{}/".format(PAO, kwat.path.clean(co), fun)

    kwat.path.make(paf)

    st = kwat.function_heat_map.make(
        ta,
        en_se_sa,
        fu,
        n_jo=SETTING["n_jo"],
        n_sa=SETTING["n_sa"],
        n_sh=SETTING["n_sh"],
        n_pl=SETTING["n_ex"],
        title="All ({})".format(fun),
        pa=paf,
        **ke,
    )

    if 0 < len(SETTING["se_"]):

        kwat.function_heat_map.make(
            ta,
            en_se_sa.reindex(labels=SETTING["se_"]),
            st,
            n_pl=None,
            title="Peek ({})".format(fun),
            **ke,
        )

        plot_peek(st.iloc[:, 0], SETTING["se_"], paf)

    for bu, se_fe_ in bu_se_fe_.items():

        pab = "{}{}/".format(paf, bu)

        kwat.path.make(pab)

        se_ = en_se_sa.index.intersection(se_fe_)

        if 0 < len(se_):

            kwat.function_heat_map.make(
                ta,
                en_se_sa.loc[se_, :],
                st,
                n_pl=SETTING["n_ex"],
                title="{} ({})".format(bu, fun),
                pa=pab,
                **ke,
            )
