==> code/11_summarize.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

bi_co_sa = pd.read_csv("{}01_comparison_sample.tsv".format(PAO), sep="\t", index_col=0)

nu_fe_sa = pd.read_csv("{}number_feature_sample.tsv".format(PAO), sep="\t", index_col=0)

en_se_sa = pd.read_csv("{}enrichment_set_sample.tsv".format(PAO), sep="\t", index_col=0)

title = "Summary"

pd.concat([nu_fe_sa, en_se_sa])

for (co, ta), bl_an_ in zip(bi_co_sa.iterrows(), SETTING["su_"]):

    ta.dropna(inplace=True)

    fu = kwat.information.get_ic

    pa = "{}summarize/{}/".format(PAO, kwat.path.clean(co))

    kwat.path.make(pa)

    bu_ = []

    bl_an_["Peek"] = SETTING["fe_"] + SETTING["se_"]

    for bl, an_ in bl_an_.items():

        bu_.append(
            {
                "name": bl,
                "data": pd.concat([nu_fe_sa, en_se_sa]).reindex(labels=an_),
                "type": "continuous",
                "statistic": read_statistic(PAO, kwat.path.clean(co), fu.__name__),
            }
        )

    kwat.function_heat_map.summarize(
        ta,
        bu_,
        ty="binary",
        st=SETTING["st"],
        title=title,
        pa="{}{}.html".format(pa, kwat.path.clean(title)),
    )
